Split public and private config/driver APIs (#6024)
