+++ redbot/core/config.py
@@ -35,7 +35,6 @@ log = logging.getLogger("red.config")
 _T = TypeVar("_T")
 
 _config_cache = weakref.WeakValueDictionary()
-_retrieved = weakref.WeakSet()
 
 
 class ConfigMeta(type):
@@ -63,14 +62,6 @@ class ConfigMeta(type):
         )
         _config_cache[key] = instance
         return instance
-
-
-def get_latest_confs() -> Tuple["Config"]:
-    global _retrieved
-    ret = set(_config_cache.values()) - set(_retrieved)
-    _retrieved |= ret
-    # noinspection PyTypeChecker
-    return tuple(ret)
 
 
 class _ValueCtxManager(Awaitable[_T], AsyncContextManager[_T]):  # pylint: disable=duplicate-bases
@@ -1527,16 +1518,6 @@ class Config(metaclass=ConfigMeta):
                 is_custom=is_custom,
             )
             return self._lock_cache.setdefault(id_data, asyncio.Lock())
-
-
-async def migrate(cur_driver_cls: Type[BaseDriver], new_driver_cls: Type[BaseDriver]) -> None:
-    """Migrate from one driver type to another."""
-    # Get custom group data
-    core_conf = Config.get_core_conf(allow_old=True)
-    core_conf.init_custom("CUSTOM_GROUPS", 2)
-    all_custom_group_data = await core_conf.custom("CUSTOM_GROUPS").all()
-
-    await cur_driver_cls.migrate_to(new_driver_cls, all_custom_group_data)
 
 
 def _str_key_dict(value: Dict[Any, _T]) -> Dict[str, _T]:
